moved uframe routes to uframe blueprints, as well as helper methods.
