+++ ooiservices/app/main/__init__.py
@@ -4,10 +4,10 @@ Main app blueprint.
 
 '''
 
-__author__ = 'M.Campbell'
+__author__ = 'M@Campbell'
 
 from flask import Blueprint
 
 api = Blueprint('main', __name__)
 
-from ooiservices.app.main import routes, authentication, user, operator_event 
+from ooiservices.app.main import routes, authentication, user, operator_event
