+++ ooiservices/app/redmine/__init__.py
@@ -1,0 +1,13 @@
+#!/usr/bin/env python
+'''
+redmine blueprint.
+
+'''
+
+__author__ = 'M@Campbell'
+
+from flask import Blueprint
+
+redmine = Blueprint('redmine', __name__)
+
+#from ooiservices.app.redmine import ...
