+++ ooiservices/app/uframe/__init__.py
@@ -1,0 +1,13 @@
+#!/usr/bin/env python
+'''
+uframe blueprint.
+
+'''
+
+__author__ = 'M@Campbell'
+
+from flask import Blueprint
+
+uframe = Blueprint('uframe', __name__)
+
+from ooiservices.app.uframe import controller
