+++ ooiservices/app/uframe/data.py
@@ -5,7 +5,7 @@ ooiservices/app/main/data.py
 Support for generating sample data
 '''
 
-
+__author__ = 'Andy Bird'
 
 import numpy as np
 import calendar
@@ -26,7 +26,7 @@ def gen_data(start_date, end_date, sampling_rate, mean, std_dev):
     x = np.arange(time0, time1, dt)
     y = np.random.normal(mean, std_dev, x.shape[0])
     xy = np.array([x,y])
-    
+
     row_order_xy = xy.T
 
 
@@ -34,6 +34,4 @@ def gen_data(start_date, end_date, sampling_rate, mean, std_dev):
     iso1 = datetime.utcfromtimestamp(time1).isoformat()
 
     return {'size' : x.shape[0], 'start_time':iso0, 'end_time':iso1, 'cols':['x','y'], 'rows':row_order_xy.tolist()}
-
-
